+++ changelog.md
@@ -16,6 +16,7 @@
   - #896 Added a cart rule for affiliates.
   - Added a button to remove the assignment of a customer to an affiliate on customer edit page.
 - Import news from RSS news feeds.
+- #971 Add a cart rule to check if the current customer is authenticated with a certain external authentication method.
 
 ### Improvements